Fix issue with hostnames that contain dots (#5)
We use the hostname when creating the csv file. When using
`with_extension` on the path we create from the hostname it will strip
the last part after a dot as it assumes that that is the file extention.

This fixes it and just creates the filename using format. This should be
fine on all platforms.