Fix newlines being saved as a valid command
When a command was received that only contained newlines it was saved as
a valid command.

This pullrequests changes that and strips ending newlines (\n and \r\n)
from the command. When that results in an empty string the command will
not be saved like any other empty command.

Also added some intigration tests for the client/server.

Fixes #21.